use correct data type gsize rather than gint when calling function
2005-06-25  Bas Driessen <bas.driessen@xobas.com>

	* providers/mysql/libmain.c
	(plugin_get_dsn_spec): use correct data type
	gsize rather than gint when calling function
	g_file_get_contents. The x86_64 platform is
	very picky about this and crashes with a
	segmentation fault.

	* providers/postgres/libmain.c
	(plugin_get_dsn_spec): idem.

	* providers/sqlite/libmain.c
	(plugin_get_dsn_spec): idem.

+++ providers/mysql/libmain.c
@@ -75,7 +75,7 @@ gchar *
 plugin_get_dsn_spec (void)
 {
 	gchar *specs, *file;
-	gint len;
+	gsize len;
 
 	file = g_build_filename (LIBGDA_DATA_DIR, "mysql_specs_dsn.xml", NULL);
 	if (g_file_get_contents (file, &specs, &len, NULL))
